BGE Shaders: Adding light include
Lighting information can now be added to a custom shader using the line
"#pragma bgl include lights". This line gets replaced with an array of
light structs. Currently only the position of the light is getting set, as
well as the number of lights. A user can adjust the default number of
maximum lights (currently 4) by defining MAX_LIGHTS before the pragma.